requirements.txt
# Telegram Bot
python-telegram-bot==20.7
aiogram==3.3.0

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
mysqlclient==2.2.0

# Technical Analysis
yfinance==0.2.28
ta==0.10.2
pandas==2.1.4
numpy==1.24.3

# Charts and Visualization
plotly==5.17.0
chart.js==4.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
cryptography==41.0.8

# Development
python-decouple==3.8 